@@ -5,97 +5,97 @@ import { loginUser } from '../../actions/authActions';
 import TextFieldGroup from '../common/TextFieldGroup';
 
 class Login extends Component {
-	constructor() {
-		super();
-		this.state = {
-			email: '',
-			password: '',
-			errors: {}
-		};
-	}
+    constructor() {
+        super();
+        this.state = {
+            email: '',
+            password: '',
+            errors: {}
+        };
+    }
 
-	onChange = e => {
-		this.setState({ [e.target.name]: e.target.value });
-	};
+    onChange = e => {
+        this.setState({ [e.target.name]: e.target.value });
+    };
 
-	onSubmit = e => {
-		e.preventDefault();
+    onSubmit = e => {
+        e.preventDefault();
 
-		const userData = {
-			email: this.state.email,
-			password: this.state.password
-		};
+        const userData = {
+            email: this.state.email,
+            password: this.state.password
+        };
 
-		this.props.loginUser(userData);
-	};
+        this.props.loginUser(userData);
+    };
 
-	componentDidMount() {
-		if (this.props.auth.isAuthenticated) {
-			this.props.history.push('/dashboard');
-		}
-	}
+    componentDidMount() {
+        if (this.props.auth.isAuthenticated) {
+            this.props.history.push('/dashboard');
+        }
+    }
 
-	UNSAFE_componentWillReceiveProps(nextProps) {
-		if (nextProps.auth.isAuthenticated) {
-			this.props.history.push('/dashboard');
-		}
+    UNSAFE_componentWillReceiveProps(nextProps) {
+        if (nextProps.auth.isAuthenticated) {
+            this.props.history.push('/dashboard');
+        }
 
-		if (nextProps.errors) {
-			this.setState({ errors: nextProps.errors });
-		}
-	}
+        if (nextProps.errors) {
+            this.setState({ errors: nextProps.errors });
+        }
+    }
 
-	render() {
-		const { errors } = this.state;
+    render() {
+        const { errors } = this.state;
 
-		return (
-			<div className="login">
-				<div className="container">
-					<div className="row">
-						<div className="col-md-8 m-auto">
-							<h1 className="display-4 text-center">Log In</h1>
-							<p className="lead text-center">
-								Sign in to your DevConnector account
-							</p>
-							<form noValidate onSubmit={this.onSubmit}>
-								<TextFieldGroup
-									placeholder="Email Address"
-									type="email"
-									onChange={this.onChange}
-									error={errors.email}
-									value={this.state.email}
-									name="email"
-								/>
-								<TextFieldGroup
-									placeholder="Password"
-									type="password"
-									onChange={this.onChange}
-									error={errors.password}
-									value={this.state.password}
-									name="password"
-								/>
-								<input type="submit" className="btn btn-info btn-block mt-4" />
-							</form>
-						</div>
-					</div>
-				</div>
-			</div>
-		);
-	}
+        return (
+            <div className="login">
+                <div className="container">
+                    <div className="row">
+                        <div className="col-md-8 m-auto">
+                            <h1 className="display-4 text-center">Log In</h1>
+                            <p className="lead text-center">
+                                Sign in to your DevConnector account
+                            </p>
+                            <form noValidate onSubmit={this.onSubmit}>
+                                <TextFieldGroup
+                                    placeholder="Email Address"
+                                    type="email"
+                                    onChange={this.onChange}
+                                    error={errors.email}
+                                    value={this.state.email}
+                                    name="email"
+                                />
+                                <TextFieldGroup
+                                    placeholder="Password"
+                                    type="password"
+                                    onChange={this.onChange}
+                                    error={errors.password}
+                                    value={this.state.password}
+                                    name="password"
+                                />
+                                <input type="submit" className="btn btn-info btn-block mt-4" />
+                            </form>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        );
+    }
 }
 
 Login.propTypes = {
-	loginUser: PropTypes.func.isRequired,
-	auth: PropTypes.object.isRequired,
-	errors: PropTypes.object.isRequired
+    loginUser: PropTypes.func.isRequired,
+    auth: PropTypes.object.isRequired,
+    errors: PropTypes.object.isRequired
 };
 
 const mapStateToProps = state => ({
-	auth: state.auth,
-	errors: state.errors
+    auth: state.auth,
+    errors: state.errors
 });
 
 export default connect(
-	mapStateToProps,
-	{ loginUser }
+    mapStateToProps,
+    { loginUser }
 )(Login);
