@@ -5,79 +5,79 @@ import TextAreaFieldGroup from '../common/TextAreaFieldGroup';
 import { addPost } from '../../actions/postActions';
 
 class PostForm extends Component {
-	constructor(props) {
-		super(props);
-		this.state = {
-			text: '',
-			errors: {}
-		};
-	}
+    constructor(props) {
+        super(props);
+        this.state = {
+            text: '',
+            errors: {}
+        };
+    }
 
-	UNSAFE_componentWillReceiveProps(newProps) {
-		if (newProps.errors) {
-			this.setState({ errors: newProps.errors });
-		}
-	}
+    UNSAFE_componentWillReceiveProps(newProps) {
+        if (newProps.errors) {
+            this.setState({ errors: newProps.errors });
+        }
+    }
 
-	onSubmit = e => {
-		e.preventDefault();
+    onSubmit = e => {
+        e.preventDefault();
 
-		const { user } = this.props.auth;
-		const newPost = {
-			text: this.state.text,
-			name: user.name,
-			avatar: user.avatar
-		};
+        const { user } = this.props.auth;
+        const newPost = {
+            text: this.state.text,
+            name: user.name,
+            avatar: user.avatar
+        };
 
-		this.props.addPost(newPost);
-		this.setState({ text: '' });
-	};
+        this.props.addPost(newPost);
+        this.setState({ text: '' });
+    };
 
-	onChange = e => {
-		this.setState({ [e.target.name]: e.target.value });
-	};
+    onChange = e => {
+        this.setState({ [e.target.name]: e.target.value });
+    };
 
-	render() {
-		const { errors } = this.state;
+    render() {
+        const { errors } = this.state;
 
-		return (
-			<div className="post-form mb-3">
-				<div className="card card-info">
-					<div className="card-header bg-info text-white">Write a post</div>
-					<div className="card-body">
-						<form onSubmit={this.onSubmit}>
-							<div className="form-group">
-								<TextAreaFieldGroup
-									placeholder="Type a message here..."
-									onChange={this.onChange}
-									value={this.state.text}
-									name="text"
-									error={errors.text}
-								/>
-							</div>
-							<button type="submit" className="btn btn-dark">
-								Submit
-							</button>
-						</form>
-					</div>
-				</div>
-			</div>
-		);
-	}
+        return (
+            <div className="post-form mb-3">
+                <div className="card card-info">
+                    <div className="card-header bg-info text-white">Write a post</div>
+                    <div className="card-body">
+                        <form onSubmit={this.onSubmit}>
+                            <div className="form-group">
+                                <TextAreaFieldGroup
+                                    placeholder="Type a message here..."
+                                    onChange={this.onChange}
+                                    value={this.state.text}
+                                    name="text"
+                                    error={errors.text}
+                                />
+                            </div>
+                            <button type="submit" className="btn btn-dark">
+                                Submit
+                            </button>
+                        </form>
+                    </div>
+                </div>
+            </div>
+        );
+    }
 }
 
 PostForm.propTypes = {
-	addPost: PropTypes.func.isRequired,
-	auth: PropTypes.object.isRequired,
-	errors: PropTypes.object.isRequired
+    addPost: PropTypes.func.isRequired,
+    auth: PropTypes.object.isRequired,
+    errors: PropTypes.object.isRequired
 };
 
 const mapStateToProps = state => ({
-	auth: state.auth,
-	errors: state.errors
+    auth: state.auth,
+    errors: state.errors
 });
 
 export default connect(
-	mapStateToProps,
-	{ addPost }
+    mapStateToProps,
+    { addPost }
 )(PostForm);
