@@ -10,70 +10,70 @@ import Spinner from '../common/Spinner';
 import { getProfileByHandle } from '../../actions/profileActions';
 
 class Profile extends Component {
-	componentDidMount() {
-		if (this.props.match.params.handle) {
-			this.props.getProfileByHandle(this.props.match.params.handle);
-		}
-	}
+    componentDidMount() {
+        if (this.props.match.params.handle) {
+            this.props.getProfileByHandle(this.props.match.params.handle);
+        }
+    }
 
-	UNSAFE_componentWillReceiveProps(nextProps) {
-		if (nextProps.profile.profile === null && this.props.profile.loading) {
-			this.props.history.push('/not-found');
-		}
-	}
+    UNSAFE_componentWillReceiveProps(nextProps) {
+        if (nextProps.profile.profile === null && this.props.profile.loading) {
+            this.props.history.push('/not-found');
+        }
+    }
 
-	render() {
-		const { profile, loading } = this.props.profile;
-		let profileContent;
+    render() {
+        const { profile, loading } = this.props.profile;
+        let profileContent;
 
-		if (profile === null || loading) {
-			profileContent = <Spinner />;
-		} else {
-			profileContent = (
-				<div>
-					<div className="row">
-						<div className="col-md-6">
-							<Link to="/profiles" className="btn btn-light mb-3 float-left">
-								Back to Profiles
-							</Link>
-						</div>
-						<div className="col-md-6" />
-					</div>
-					<ProfileHeader profile={profile} />
-					<ProfileAbout profile={profile} />
-					<ProfileCreds
-						education={profile.education}
-						experience={profile.experience}
-					/>
-					{profile.githubusername ? (
-						<ProfileGithub username={profile.githubusername} />
-					) : null}
-				</div>
-			);
-		}
+        if (profile === null || loading) {
+            profileContent = <Spinner />;
+        } else {
+            profileContent = (
+                <div>
+                    <div className="row">
+                        <div className="col-md-6">
+                            <Link to="/profiles" className="btn btn-light mb-3 float-left">
+                                Back to Profiles
+                            </Link>
+                        </div>
+                        <div className="col-md-6" />
+                    </div>
+                    <ProfileHeader profile={profile} />
+                    <ProfileAbout profile={profile} />
+                    <ProfileCreds
+                        education={profile.education}
+                        experience={profile.experience}
+                    />
+                    {profile.githubusername ? (
+                        <ProfileGithub username={profile.githubusername} />
+                    ) : null}
+                </div>
+            );
+        }
 
-		return (
-			<div className="profile">
-				<div className="container">
-					<div className="row">
-						<div className="col-md-12">{profileContent}</div>
-					</div>
-				</div>
-			</div>
-		);
-	}
+        return (
+            <div className="profile">
+                <div className="container">
+                    <div className="row">
+                        <div className="col-md-12">{profileContent}</div>
+                    </div>
+                </div>
+            </div>
+        );
+    }
 }
 
 Profile.propTypes = {
-	profile: PropTypes.object.isRequired,
-	getProfileByHandle: PropTypes.func.isRequired
+    profile: PropTypes.object.isRequired,
+    getProfileByHandle: PropTypes.func.isRequired
 };
 
 const mapStateToProps = state => ({
-	profile: state.profile
+    profile: state.profile
 });
 
 export default connect(
-	mapStateToProps,
-	{ getProfileByHandle }
+    mapStateToProps,
+    { getProfileByHandle }
 )(Profile);
