@@ -4,7 +4,7 @@ import { Provider } from 'react-redux';
 import store from './store';
 import jwt_decode from 'jwt-decode';
 import setAuthToken from './utils/setAuthToken';
-import { setCurrentUser, logoutUser } from './actions/authActions';
+import { logoutUser, setCurrentUser } from './actions/authActions';
 import { clearCurrentProfile } from './actions/profileActions';
 import PrivateRoute from './components/common/PrivateRoute';
 
@@ -29,83 +29,83 @@ import './App.css';
 // Check for token
 var token = localStorage.jwtToken;
 if (token) {
-	// Set auth token header auth
-	setAuthToken(token);
-	// Decode token and get user info and expir
-	const decoded = jwt_decode(token);
-	// Set user and isAuthenticated
-	store.dispatch(setCurrentUser(decoded));
+    // Set auth token header auth
+    setAuthToken(token);
+    // Decode token and get user info and expir
+    const decoded = jwt_decode(token);
+    // Set user and isAuthenticated
+    store.dispatch(setCurrentUser(decoded));
 
-	// Check for expired token
-	const currentTime = Date.now() / 1000;
-	if (decoded.exp < currentTime) {
-		// Logout the user
-		store.dispatch(logoutUser());
-		// Clear current profile
-		store.dispatch(clearCurrentProfile());
-		// // Redirect to login
-		// window.location.href = '/login';
-	}
+    // Check for expired token
+    const currentTime = Date.now() / 1000;
+    if (decoded.exp < currentTime) {
+        // Logout the user
+        store.dispatch(logoutUser());
+        // Clear current profile
+        store.dispatch(clearCurrentProfile());
+        // // Redirect to login
+        // window.location.href = '/login';
+    }
 }
 
 class App extends Component {
-	render() {
-		return (
-			<Provider store={store}>
-				<Router>
-					<div className="App">
-						<Navbar />
-						<Route exact path="/" component={Landing} />
-						<div className="container">
-							<Route exact path="/register" component={Register} />
-							<Route exact path="/login" component={Login} />
-							<Route exact path="/profiles" component={Profiles} />
-							<Route exact path="/profile/:handle" component={Profile} />
-							<Switch>
-								<PrivateRoute exact path="/dashboard" component={Dashboard} />
-							</Switch>
-							<Switch>
-								<PrivateRoute
-									exact
-									path="/create-profile"
-									component={CreateProfile}
-								/>
-							</Switch>
-							<Switch>
-								<PrivateRoute
-									exact
-									path="/edit-profile"
-									component={EditProfile}
-								/>
-							</Switch>
-							<Switch>
-								<PrivateRoute
-									exact
-									path="/add-experience"
-									component={AddExperience}
-								/>
-							</Switch>
-							<Switch>
-								<PrivateRoute
-									exact
-									path="/add-education"
-									component={AddEducation}
-								/>
-							</Switch>
-							<Switch>
-								<PrivateRoute exact path="/feed" component={Posts} />
-							</Switch>
-							<Switch>
-								<PrivateRoute exact path="/post/:id" component={Post} />
-							</Switch>
-							<Route exact path="/not-found" component={NotFound} />
-						</div>
-						<Footer />
-					</div>
-				</Router>
-			</Provider>
-		);
-	}
+    render() {
+        return (
+            <Provider store={store}>
+                <Router>
+                    <div className="App">
+                        <Navbar />
+                        <Route exact path="/" component={Landing} />
+                        <div className="container">
+                            <Route exact path="/register" component={Register} />
+                            <Route exact path="/login" component={Login} />
+                            <Route exact path="/profiles" component={Profiles} />
+                            <Route exact path="/profile/:handle" component={Profile} />
+                            <Switch>
+                                <PrivateRoute exact path="/dashboard" component={Dashboard} />
+                            </Switch>
+                            <Switch>
+                                <PrivateRoute
+                                    exact
+                                    path="/create-profile"
+                                    component={CreateProfile}
+                                />
+                            </Switch>
+                            <Switch>
+                                <PrivateRoute
+                                    exact
+                                    path="/edit-profile"
+                                    component={EditProfile}
+                                />
+                            </Switch>
+                            <Switch>
+                                <PrivateRoute
+                                    exact
+                                    path="/add-experience"
+                                    component={AddExperience}
+                                />
+                            </Switch>
+                            <Switch>
+                                <PrivateRoute
+                                    exact
+                                    path="/add-education"
+                                    component={AddEducation}
+                                />
+                            </Switch>
+                            <Switch>
+                                <PrivateRoute exact path="/feed" component={Posts} />
+                            </Switch>
+                            <Switch>
+                                <PrivateRoute exact path="/post/:id" component={Post} />
+                            </Switch>
+                            <Route exact path="/not-found" component={NotFound} />
+                        </div>
+                        <Footer />
+                    </div>
+                </Router>
+            </Provider>
+        );
+    }
 }
 
 export default App;
