@@ -7,55 +7,55 @@ import { GET_ERRORS, SET_CURRENT_USER } from './types';
 // We want to wait for the response from the server and then dispatch the action
 // Redux thunk allows us to use dispatch inside an action creator
 export const registerUser = (userData, history) => dispatch => {
-	axios
-		.post('/api/users/register', userData)
-		.then(res => history.push('/login'))
-		.catch(err =>
-			dispatch({
-				type: GET_ERRORS,
-				payload: err.response.data
-			})
-		);
+    axios
+        .post('/api/users/register', userData)
+        .then(() => history.push('/login'))
+        .catch(err =>
+            dispatch({
+                type: GET_ERRORS,
+                payload: err.response.data
+            })
+        );
 };
 
 // Login - Get User Token
 export const loginUser = userData => dispatch => {
-	axios
-		.post('/api/users/login', userData)
-		.then(res => {
-			// Save to localStorage
-			const { token } = res.data;
-			// Set token to localStorage
-			localStorage.setItem('jwtToken', token);
-			// Set token to Auth header
-			setAuthToken(token);
-			// Decode token to get user data
-			const decoded = jwt_decode(token);
-			// Set current user
-			dispatch(setCurrentUser(decoded));
-		})
-		.catch(err => {
-			dispatch({
-				type: GET_ERRORS,
-				payload: err.response.data
-			});
-		});
+    axios
+        .post('/api/users/login', userData)
+        .then(res => {
+            // Save to localStorage
+            const { token } = res.data;
+            // Set token to localStorage
+            localStorage.setItem('jwtToken', token);
+            // Set token to Auth header
+            setAuthToken(token);
+            // Decode token to get user data
+            const decoded = jwt_decode(token);
+            // Set current user
+            dispatch(setCurrentUser(decoded));
+        })
+        .catch(err => {
+            dispatch({
+                type: GET_ERRORS,
+                payload: err.response.data
+            });
+        });
 };
 
 // Set logged in user
 export const setCurrentUser = decoded => {
-	return {
-		type: SET_CURRENT_USER,
-		payload: decoded
-	};
+    return {
+        type: SET_CURRENT_USER,
+        payload: decoded
+    };
 };
 
 // Log user out
 export const logoutUser = () => dispatch => {
-	// Remove token from localStorage
-	localStorage.removeItem('jwtToken');
-	// Remove auth header for future requests
-	setAuthToken(false);
-	// Set current user to {} which will set isAuthenticated to false
-	dispatch(setCurrentUser({}));
+    // Remove token from localStorage
+    localStorage.removeItem('jwtToken');
+    // Remove auth header for future requests
+    setAuthToken(false);
+    // Set current user to {} which will set isAuthenticated to false
+    dispatch(setCurrentUser({}));
 };
