@@ -1,166 +1,158 @@
 import axios from 'axios';
-import {
-	ADD_POST,
-	GET_ERRORS,
-	GET_POSTS,
-	POST_LOADING,
-	DELETE_POST,
-	GET_POST,
-	CLEAR_ERRORS
-} from './types';
+import { ADD_POST, CLEAR_ERRORS, DELETE_POST, GET_ERRORS, GET_POST, GET_POSTS, POST_LOADING } from './types';
 
 // Add Posts
 export const addPost = postData => dispatch => {
-	dispatch(clearErrors());
+    dispatch(clearErrors());
 
-	axios
-		.post('/api/posts', postData)
-		.then(res =>
-			dispatch({
-				type: ADD_POST,
-				payload: res.data
-			})
-		)
-		.catch(err =>
-			dispatch({
-				type: GET_ERRORS,
-				payload: err.response.data
-			})
-		);
+    axios
+        .post('/api/posts', postData)
+        .then(res =>
+            dispatch({
+                type: ADD_POST,
+                payload: res.data
+            })
+        )
+        .catch(err =>
+            dispatch({
+                type: GET_ERRORS,
+                payload: err.response.data
+            })
+        );
 };
 
 // Get Posts
 export const getPosts = () => dispatch => {
-	dispatch(setPostLoading());
+    dispatch(setPostLoading());
 
-	axios
-		.get('/api/posts')
-		.then(res =>
-			dispatch({
-				type: GET_POSTS,
-				payload: res.data
-			})
-		)
-		.catch(err =>
-			dispatch({
-				type: GET_POSTS,
-				payload: null
-			})
-		);
+    axios
+        .get('/api/posts')
+        .then(res =>
+            dispatch({
+                type: GET_POSTS,
+                payload: res.data
+            })
+        )
+        .catch(() =>
+            dispatch({
+                type: GET_POSTS,
+                payload: null
+            })
+        );
 };
 
 // Get Post
 export const getPost = id => dispatch => {
-	dispatch(setPostLoading());
+    dispatch(setPostLoading());
 
-	axios
-		.get(`/api/posts/${id}`)
-		.then(res =>
-			dispatch({
-				type: GET_POST,
-				payload: res.data
-			})
-		)
-		.catch(err =>
-			dispatch({
-				type: GET_POST,
-				payload: null
-			})
-		);
+    axios
+        .get(`/api/posts/${id}`)
+        .then(res =>
+            dispatch({
+                type: GET_POST,
+                payload: res.data
+            })
+        )
+        .catch(() =>
+            dispatch({
+                type: GET_POST,
+                payload: null
+            })
+        );
 };
 
 // Delete Post
 export const deletePost = id => dispatch => {
-	axios
-		.delete(`/api/posts/${id}`)
-		.then(res =>
-			dispatch({
-				type: DELETE_POST,
-				payload: id
-			})
-		)
-		.catch(err =>
-			dispatch({
-				type: GET_ERRORS,
-				payload: err.response.data
-			})
-		);
+    axios
+        .delete(`/api/posts/${id}`)
+        .then(() =>
+            dispatch({
+                type: DELETE_POST,
+                payload: id
+            })
+        )
+        .catch(err =>
+            dispatch({
+                type: GET_ERRORS,
+                payload: err.response.data
+            })
+        );
 };
 
 //Add Like
 export const addLike = id => dispatch => {
-	axios
-		.post(`/api/posts/like/${id}`)
-		.then(res => dispatch(getPosts()))
-		.catch(err =>
-			dispatch({
-				type: GET_ERRORS,
-				payload: err.response.data
-			})
-		);
+    axios
+        .post(`/api/posts/like/${id}`)
+        .then(() => dispatch(getPosts()))
+        .catch(err =>
+            dispatch({
+                type: GET_ERRORS,
+                payload: err.response.data
+            })
+        );
 };
 
 //Remove Like
 export const removeLike = id => dispatch => {
-	axios
-		.post(`/api/posts/unlike/${id}`)
-		.then(res => dispatch(getPosts()))
-		.catch(err =>
-			dispatch({
-				type: GET_ERRORS,
-				payload: err.response.data
-			})
-		);
+    axios
+        .post(`/api/posts/unlike/${id}`)
+        .then(() => dispatch(getPosts()))
+        .catch(err =>
+            dispatch({
+                type: GET_ERRORS,
+                payload: err.response.data
+            })
+        );
 };
 
 // Add Comment
 export const addComment = (postId, commentData) => dispatch => {
-	dispatch(clearErrors());
+    dispatch(clearErrors());
 
-	axios
-		.post(`/api/posts/comment/${postId}`, commentData)
-		.then(res =>
-			dispatch({
-				type: GET_POST,
-				payload: res.data
-			})
-		)
-		.catch(err =>
-			dispatch({
-				type: GET_ERRORS,
-				payload: err.response.data
-			})
-		);
+    axios
+        .post(`/api/posts/comment/${postId}`, commentData)
+        .then(res =>
+            dispatch({
+                type: GET_POST,
+                payload: res.data
+            })
+        )
+        .catch(err =>
+            dispatch({
+                type: GET_ERRORS,
+                payload: err.response.data
+            })
+        );
 };
 
 // Delete Comment
 export const deleteComment = (postId, commentId) => dispatch => {
-	axios
-		.delete(`/api/posts/comment/${postId}/${commentId}`)
-		.then(res =>
-			dispatch({
-				type: GET_POST,
-				payload: res.data
-			})
-		)
-		.catch(err =>
-			dispatch({
-				type: GET_ERRORS,
-				payload: err.response.data
-			})
-		);
+    axios
+        .delete(`/api/posts/comment/${postId}/${commentId}`)
+        .then(res =>
+            dispatch({
+                type: GET_POST,
+                payload: res.data
+            })
+        )
+        .catch(err =>
+            dispatch({
+                type: GET_ERRORS,
+                payload: err.response.data
+            })
+        );
 };
 
 // Set loading state
 export const setPostLoading = () => {
-	return {
-		type: POST_LOADING
-	};
+    return {
+        type: POST_LOADING
+    };
 };
 
 // Clear Errors
 export const clearErrors = () => {
-	return {
-		type: CLEAR_ERRORS
-	};
+    return {
+        type: CLEAR_ERRORS
+    };
 };
