@@ -4,47 +4,46 @@ import PropTypes from 'prop-types';
 
 // Pass properties in the functional (stateless) component
 const InputGroup = ({
-	name,
-	placeholder,
-	value,
-	error,
-	icon,
-	type,
-	onChange
+    name,
+    placeholder,
+    value,
+    error,
+    icon,
+    onChange
 }) => {
-	return (
-		<div className="input-group mb-3">
-			<div className="input-group-prepend">
-				<span className="input-group-text">
-					<i className={icon} />
-				</span>
-			</div>
-			<input
-				className={classnames('form-control form-control-lg', {
-					'is-invalid': error
-				})}
-				placeholder={placeholder}
-				value={value}
-				onChange={onChange}
-				name={name}
-			/>
-			{error && <div className="invalid-feedback">{error}</div>}
-		</div>
-	);
+    return (
+        <div className="input-group mb-3">
+            <div className="input-group-prepend">
+                <span className="input-group-text">
+                    <i className={icon} />
+                </span>
+            </div>
+            <input
+                className={classnames('form-control form-control-lg', {
+                    'is-invalid': error
+                })}
+                placeholder={placeholder}
+                value={value}
+                onChange={onChange}
+                name={name}
+            />
+            {error && <div className="invalid-feedback">{error}</div>}
+        </div>
+    );
 };
 
 InputGroup.propTypes = {
-	name: PropTypes.string.isRequired,
-	placeholder: PropTypes.string,
-	value: PropTypes.string.isRequired,
-	icon: PropTypes.string,
-	error: PropTypes.string,
-	type: PropTypes.string.isRequired,
-	onChange: PropTypes.func.isRequired
+    name: PropTypes.string.isRequired,
+    placeholder: PropTypes.string,
+    value: PropTypes.string.isRequired,
+    icon: PropTypes.string,
+    error: PropTypes.string,
+    type: PropTypes.string.isRequired,
+    onChange: PropTypes.func.isRequired
 };
 
 InputGroup.defaultProps = {
-	type: 'text'
+    type: 'text'
 };
 
 export default InputGroup;
