@@ -1,28 +1,28 @@
 import React from 'react';
-import { Route, Redirect } from 'react-router-dom';
+import { Redirect, Route } from 'react-router-dom';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
 // This is a wrapper around the Route, which checks for authentication
 const PrivateRoute = ({ component: Component, auth, ...rest }) => (
-	<Route
-		{...rest}
-		render={props =>
-			auth.isAuthenticated === true ? (
-				<Component {...props} />
-			) : (
-					<Redirect to="/" />
-				)
-		}
-	/>
+    <Route
+        {...rest}
+        render={props =>
+            auth.isAuthenticated === true ? (
+                <Component {...props} />
+            ) : (
+                    <Redirect to="/" />
+                )
+        }
+    />
 );
 
 PrivateRoute.propTypes = {
-	auth: PropTypes.object.isRequired
+    auth: PropTypes.object.isRequired
 };
 
 const mapStateToProps = state => ({
-	auth: state.auth
+    auth: state.auth
 });
 
 export default connect(mapStateToProps)(PrivateRoute);
