@@ -4,37 +4,37 @@ import PropTypes from 'prop-types';
 
 // Pass properties in the functional (stateless) component
 const SelectListGroup = ({ name, value, error, info, onChange, options }) => {
-	const selectOptions = options.map(option => (
-		<option key={option.label} value={option.value}>
-			{option.label}
-		</option>
-	));
+    const selectOptions = options.map(option => (
+        <option key={option.label} value={option.value}>
+            {option.label}
+        </option>
+    ));
 
-	return (
-		<div className="form-group">
-			<select
-				className={classnames('form-control form-control-lg', {
-					'is-invalid': error
-				})}
-				value={value}
-				onChange={onChange}
-				name={name}
-			>
-				{selectOptions}
-			</select>
-			{info && <small className="form-text text-muted">{info}</small>}
-			{error && <div className="invalid-feedback">{error}</div>}
-		</div>
-	);
+    return (
+        <div className="form-group">
+            <select
+                className={classnames('form-control form-control-lg', {
+                    'is-invalid': error
+                })}
+                value={value}
+                onChange={onChange}
+                name={name}
+            >
+                {selectOptions}
+            </select>
+            {info && <small className="form-text text-muted">{info}</small>}
+            {error && <div className="invalid-feedback">{error}</div>}
+        </div>
+    );
 };
 
 SelectListGroup.propTypes = {
-	name: PropTypes.string.isRequired,
-	value: PropTypes.string.isRequired,
-	info: PropTypes.string,
-	error: PropTypes.string,
-	onChange: PropTypes.func.isRequired,
-	options: PropTypes.array.isRequired
+    name: PropTypes.string.isRequired,
+    value: PropTypes.string.isRequired,
+    info: PropTypes.string,
+    error: PropTypes.string,
+    onChange: PropTypes.func.isRequired,
+    options: PropTypes.array.isRequired
 };
 
 export default SelectListGroup;
