@@ -6,106 +6,106 @@ import { logoutUser } from '../../actions/authActions';
 import { clearCurrentProfile } from '../../actions/profileActions';
 
 class Navbar extends Component {
-	onLogoutClick = e => {
-		e.preventDefault();
-		this.props.clearCurrentProfile();
-		this.props.logoutUser();
-		// Cannot use history here, bc this is not a Route
-		window.location.href = '/';
-	};
+    onLogoutClick = e => {
+        e.preventDefault();
+        this.props.clearCurrentProfile();
+        this.props.logoutUser();
+        // Cannot use history here, bc this is not a Route
+        window.location.href = '/';
+    };
 
-	render() {
-		const { isAuthenticated, user } = this.props.auth;
+    render() {
+        const { isAuthenticated, user } = this.props.auth;
 
-		const authLinks = (
-			<ul className="navbar-nav ml-auto">
-				<li className="nav-item">
-					<Link className="nav-link" to="/feed">
-						Post Feed
-					</Link>
-				</li>
-				<li className="nav-item">
-					<Link className="nav-link" to="/dashboard">
-						Dashboard
-					</Link>
-				</li>
-				<li className="nav-item">
-					<Link
-						onClick={this.onLogoutClick}
-						className="nav-link"
-						to="/dashboard"
-					>
-						<img
-							src={user.avatar}
-							alt={user.name}
-							className="rounded-circle"
-							title="You must have a Gravatar connected to your email to display an image"
-							style={{ width: '25px', marginRight: '5px' }}
-						/>
-						Logout
-					</Link>
-				</li>
-			</ul>
-		);
+        const authLinks = (
+            <ul className="navbar-nav ml-auto">
+                <li className="nav-item">
+                    <Link className="nav-link" to="/feed">
+                        Post Feed
+                    </Link>
+                </li>
+                <li className="nav-item">
+                    <Link className="nav-link" to="/dashboard">
+                        Dashboard
+                    </Link>
+                </li>
+                <li className="nav-item">
+                    <Link
+                        onClick={this.onLogoutClick}
+                        className="nav-link"
+                        to="/dashboard"
+                    >
+                        <img
+                            src={user.avatar}
+                            alt={user.name}
+                            className="rounded-circle"
+                            title="You must have a Gravatar connected to your email to display an image"
+                            style={{ width: '25px', marginRight: '5px' }}
+                        />
+                        Logout
+                    </Link>
+                </li>
+            </ul>
+        );
 
-		const guestLinks = (
-			<ul className="navbar-nav ml-auto">
-				<li className="nav-item">
-					<Link className="nav-link" to="/register">
-						Sign Up
-					</Link>
-				</li>
-				<li className="nav-item">
-					<Link className="nav-link" to="/login">
-						Login
-					</Link>
-				</li>
-			</ul>
-		);
+        const guestLinks = (
+            <ul className="navbar-nav ml-auto">
+                <li className="nav-item">
+                    <Link className="nav-link" to="/register">
+                        Sign Up
+                    </Link>
+                </li>
+                <li className="nav-item">
+                    <Link className="nav-link" to="/login">
+                        Login
+                    </Link>
+                </li>
+            </ul>
+        );
 
-		// Return is only for JSX
-		return (
-			<nav className="navbar navbar-expand-sm navbar-dark bg-dark mb-4 sticky-top">
-				<div className="container">
-					<Link className="navbar-brand" to="/">
-						DevConnector
-					</Link>
-					<button
-						className="navbar-toggler"
-						type="button"
-						data-toggle="collapse"
-						data-target="#mobile-nav"
-					>
-						<span className="navbar-toggler-icon" />
-					</button>
+        // Return is only for JSX
+        return (
+            <nav className="navbar navbar-expand-sm navbar-dark bg-dark mb-4 sticky-top">
+                <div className="container">
+                    <Link className="navbar-brand" to="/">
+                        DevConnector
+                    </Link>
+                    <button
+                        className="navbar-toggler"
+                        type="button"
+                        data-toggle="collapse"
+                        data-target="#mobile-nav"
+                    >
+                        <span className="navbar-toggler-icon" />
+                    </button>
 
-					<div className="collapse navbar-collapse" id="mobile-nav">
-						<ul className="navbar-nav mr-auto">
-							<li className="nav-item">
-								<Link className="nav-link" to="/profiles">
-									{' '}
-									Developers
-								</Link>
-							</li>
-						</ul>
-						{isAuthenticated ? authLinks : guestLinks}
-					</div>
-				</div>
-			</nav>
-		);
-	}
+                    <div className="collapse navbar-collapse" id="mobile-nav">
+                        <ul className="navbar-nav mr-auto">
+                            <li className="nav-item">
+                                <Link className="nav-link" to="/profiles">
+                                    {' '}
+                                    Developers
+                                </Link>
+                            </li>
+                        </ul>
+                        {isAuthenticated ? authLinks : guestLinks}
+                    </div>
+                </div>
+            </nav>
+        );
+    }
 }
 
 Navbar.propTypes = {
-	logoutUser: PropTypes.func.isRequired,
-	auth: PropTypes.object.isRequired
+    logoutUser: PropTypes.func.isRequired,
+    auth: PropTypes.object.isRequired
 };
 
 const mapStateToProps = state => ({
-	auth: state.auth
+    auth: state.auth
 });
 
 export default connect(
-	mapStateToProps,
-	{ logoutUser, clearCurrentProfile }
+    mapStateToProps,
+    { logoutUser, clearCurrentProfile }
 )(Navbar);
