@@ -1,14 +1,14 @@
 import React, { Component } from 'react';
 
 class NotFound extends Component {
-	render() {
-		return (
-			<div>
-				<h1 className="display-4">Page Not Found</h1>
-				<p>This page does not exist</p>
-			</div>
-		);
-	}
+    render() {
+        return (
+            <div>
+                <h1 className="display-4">Page Not Found</h1>
+                <p>This page does not exist</p>
+            </div>
+        );
+    }
 }
 
 export default NotFound;
