@@ -3,18 +3,18 @@ import PropTypes from 'prop-types';
 import CommentItem from './CommentItem';
 
 class CommentFeed extends Component {
-	render() {
-		const { comments, postId } = this.props;
+    render() {
+        const { comments, postId } = this.props;
 
-		return comments.map(comment => (
-			<CommentItem key={comment._id} comment={comment} postId={postId} />
-		));
-	}
+        return comments.map(comment => (
+            <CommentItem key={comment._id} comment={comment} postId={postId} />
+        ));
+    }
 }
 
 CommentFeed.propTypes = {
-	comments: PropTypes.array.isRequired,
-	postId: PropTypes.string.isRequired
+    comments: PropTypes.array.isRequired,
+    postId: PropTypes.string.isRequired
 };
 
 export default CommentFeed;
