@@ -9,53 +9,53 @@ import Spinner from '../common/Spinner';
 import { getPost } from '../../actions/postActions';
 
 class Post extends Component {
-	componentDidMount() {
-		this.props.getPost(this.props.match.params.id);
-	}
+    componentDidMount() {
+        this.props.getPost(this.props.match.params.id);
+    }
 
-	render() {
-		const { post, loading } = this.props.post;
-		let postContent;
+    render() {
+        const { post, loading } = this.props.post;
+        let postContent;
 
-		if (post === null || loading || Object.keys(post).length === 0) {
-			postContent = <Spinner />;
-		} else {
-			postContent = (
-				<div>
-					<PostItem post={post} showActions={false} />
-					<CommentForm postId={post._id} />
-					<CommentFeed comments={post.comments} postId={post._id} />
-				</div>
-			);
-		}
+        if (post === null || loading || Object.keys(post).length === 0) {
+            postContent = <Spinner />;
+        } else {
+            postContent = (
+                <div>
+                    <PostItem post={post} showActions={false} />
+                    <CommentForm postId={post._id} />
+                    <CommentFeed comments={post.comments} postId={post._id} />
+                </div>
+            );
+        }
 
-		return (
-			<div className="post">
-				<div className="container">
-					<div className="row">
-						<div className="col-md-12">
-							<Link to="/feed" className="btn btn-light mb-3">
-								Back to Feed
-							</Link>
-							{postContent}
-						</div>
-					</div>
-				</div>
-			</div>
-		);
-	}
+        return (
+            <div className="post">
+                <div className="container">
+                    <div className="row">
+                        <div className="col-md-12">
+                            <Link to="/feed" className="btn btn-light mb-3">
+                                Back to Feed
+                            </Link>
+                            {postContent}
+                        </div>
+                    </div>
+                </div>
+            </div>
+        );
+    }
 }
 
 Post.propTypes = {
-	getPost: PropTypes.func.isRequired,
-	post: PropTypes.object.isRequired
+    getPost: PropTypes.func.isRequired,
+    post: PropTypes.object.isRequired
 };
 
 const mapStateToProps = state => ({
-	post: state.post
+    post: state.post
 });
 
 export default connect(
-	mapStateToProps,
-	{ getPost }
+    mapStateToProps,
+    { getPost }
 )(Post);
