@@ -3,15 +3,15 @@ import PropTypes from 'prop-types';
 import PostItem from './PostItem';
 
 class PostFeed extends Component {
-	render() {
-		const { posts } = this.props;
+    render() {
+        const { posts } = this.props;
 
-		return posts.map(post => <PostItem key={post._id} post={post} />);
-	}
+        return posts.map(post => <PostItem key={post._id} post={post} />);
+    }
 }
 
 PostFeed.propTypes = {
-	posts: PropTypes.array.isRequired
+    posts: PropTypes.array.isRequired
 };
 
 export default PostFeed;
