@@ -3,111 +3,111 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import classNames from 'classnames';
 import { Link } from 'react-router-dom';
-import { deletePost, addLike, removeLike } from '../../actions/postActions';
+import { addLike, deletePost, removeLike } from '../../actions/postActions';
 
 class PostItem extends Component {
-	onDeleteClick = (event, id) => {
-		event.preventDefault();
-		this.props.deletePost(id);
-	};
+    onDeleteClick = (event, id) => {
+        event.preventDefault();
+        this.props.deletePost(id);
+    };
 
-	onLikeClick = (event, id) => {
-		event.preventDefault();
-		this.props.addLike(id);
-	};
+    onLikeClick = (event, id) => {
+        event.preventDefault();
+        this.props.addLike(id);
+    };
 
-	onUnlikeClick = (event, id) => {
-		event.preventDefault();
-		this.props.removeLike(id);
-	};
+    onUnlikeClick = (event, id) => {
+        event.preventDefault();
+        this.props.removeLike(id);
+    };
 
-	findUserLike(likes) {
-		const { auth } = this.props;
-		if (likes.filter(like => like.user === auth.user.id).length > 0) {
-			return true;
-		} else {
-			return false;
-		}
-	}
+    findUserLike(likes) {
+        const { auth } = this.props;
+        if (likes.filter(like => like.user === auth.user.id).length > 0) {
+            return true;
+        } else {
+            return false;
+        }
+    }
 
-	render() {
-		const { post, auth, showActions } = this.props;
+    render() {
+        const { post, auth, showActions } = this.props;
 
-		return (
-			<div className="card card-body mb-3">
-				<div className="row">
-					<div className="col-md-2">
-						<a href="profile.html">
-							<img
-								className="rounded-circle d-none d-md-block"
-								src={post.avatar}
-								alt=""
-							/>
-						</a>
-						<br />
-						<p className="text-center">{post.name}</p>
-					</div>
-					<div className="col-md-10">
-						<p className="lead">{post.text}</p>
-						{showActions ? (
-							<span>
-								<button
-									type="button"
-									onClick={event => this.onLikeClick(event, post._id)}
-									className="btn btn-light mr-1"
-								>
-									<i
-										className={classNames('fas fa-thumbs-up', {
-											'text-info': this.findUserLike(post.likes)
-										})}
-									/>
-									<span className="badge badge-light">{post.likes.length}</span>
-								</button>
-								<button
-									type="button"
-									onClick={event => this.onUnlikeClick(event, post._id)}
-									className="btn btn-light mr-1"
-								>
-									<i className="text-secondary fas fa-thumbs-down" />
-								</button>
-								<Link to={`/post/${post._id}`} className="btn btn-info mr-1">
-									Comments
+        return (
+            <div className="card card-body mb-3">
+                <div className="row">
+                    <div className="col-md-2">
+                        <a href="profile.html">
+                            <img
+                                className="rounded-circle d-none d-md-block"
+                                src={post.avatar}
+                                alt=""
+                            />
+                        </a>
+                        <br />
+                        <p className="text-center">{post.name}</p>
+                    </div>
+                    <div className="col-md-10">
+                        <p className="lead">{post.text}</p>
+                        {showActions ? (
+                            <span>
+                                <button
+                                    type="button"
+                                    onClick={event => this.onLikeClick(event, post._id)}
+                                    className="btn btn-light mr-1"
+                                >
+                                    <i
+                                        className={classNames('fas fa-thumbs-up', {
+                                            'text-info': this.findUserLike(post.likes)
+                                        })}
+                                    />
+                                    <span className="badge badge-light">{post.likes.length}</span>
+                                </button>
+                                <button
+                                    type="button"
+                                    onClick={event => this.onUnlikeClick(event, post._id)}
+                                    className="btn btn-light mr-1"
+                                >
+                                    <i className="text-secondary fas fa-thumbs-down" />
+                                </button>
+                                <Link to={`/post/${post._id}`} className="btn btn-info mr-1">
+                                    Comments
 								</Link>
-								{post.user === auth.user.id ? (
-									<button
-										className="btn btn-danger mr-1"
-										type="button"
-										onClick={event => this.onDeleteClick(event, post._id)}
-									>
-										<i className="fas fa-times" />
-									</button>
-								) : null}
-							</span>
-						) : null}
-					</div>
-				</div>
-			</div>
-		);
-	}
+                                {post.user === auth.user.id ? (
+                                    <button
+                                        className="btn btn-danger mr-1"
+                                        type="button"
+                                        onClick={event => this.onDeleteClick(event, post._id)}
+                                    >
+                                        <i className="fas fa-times" />
+                                    </button>
+                                ) : null}
+                            </span>
+                        ) : null}
+                    </div>
+                </div>
+            </div>
+        );
+    }
 }
 
 PostItem.defaultProps = {
-	showActions: true
+    showActions: true
 };
 
 PostItem.propTypes = {
-	post: PropTypes.object.isRequired,
-	auth: PropTypes.object.isRequired,
-	deletePost: PropTypes.func.isRequired,
-	addLike: PropTypes.func.isRequired,
-	removeLike: PropTypes.func.isRequired
+    post: PropTypes.object.isRequired,
+    auth: PropTypes.object.isRequired,
+    deletePost: PropTypes.func.isRequired,
+    addLike: PropTypes.func.isRequired,
+    removeLike: PropTypes.func.isRequired
 };
 
 const mapStateToProps = state => ({
-	auth: state.auth
+    auth: state.auth
 });
 
 export default connect(
-	mapStateToProps,
-	{ deletePost, addLike, removeLike }
+    mapStateToProps,
+    { deletePost, addLike, removeLike }
 )(PostItem);
