@@ -1,14 +1,14 @@
-import { GET_ERRORS, CLEAR_ERRORS } from '../actions/types';
+import { CLEAR_ERRORS, GET_ERRORS } from '../actions/types';
 
 const initialState = {};
 
 export default function (state = initialState, action) {
-	switch (action.type) {
-		case GET_ERRORS:
-			return action.payload;
-		case CLEAR_ERRORS:
-			return {};
-		default:
-			return state;
-	}
+    switch (action.type) {
+        case GET_ERRORS:
+            return action.payload;
+        case CLEAR_ERRORS:
+            return {};
+        default:
+            return state;
+    }
 }
