@@ -1,4 +1,4 @@
-import { createStore, applyMiddleware } from 'redux';
+import { applyMiddleware, createStore } from 'redux';
 import thunk from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension/logOnlyInProduction'; // Use the npm package to avoid errors on mobile browsers
 import rootReducer from './reducers'; // We don't need to specify index.js here as it is the default
@@ -8,9 +8,9 @@ const initialState = {};
 const middleware = [thunk];
 
 const store = createStore(
-	rootReducer,
-	initialState,
-	composeWithDevTools(applyMiddleware(...middleware))
+    rootReducer,
+    initialState,
+    composeWithDevTools(applyMiddleware(...middleware))
 );
 
 export default store;
