@@ -1,13 +1,13 @@
 import axios from 'axios';
 
 const setAuthToken = token => {
-	if (token) {
-		// Apply to every request
-		axios.defaults.headers.common['Authorization'] = token;
-	} else {
-		// Delete auth header
-		delete axios.defaults.headers.common['Authorization'];
-	}
+    if (token) {
+        // Apply to every request
+        axios.defaults.headers.common['Authorization'] = token;
+    } else {
+        // Delete auth header
+        delete axios.defaults.headers.common['Authorization'];
+    }
 };
 
 export default setAuthToken;
