@@ -1,6 +1,6 @@
 // These credentials were created only for development purposes
 module.exports = {
-	oldMLABMongoURI: 'mongodb://theo:tsoba405012@ds145750.mlab.com:45750/dev_connector',
-	mongoURI: 'mongodb+srv://theo:tsoba405012@dev-connector.bytbr.mongodb.net/dev_connector?retryWrites=true&w=majority',
-	secretOrKey: 'whatever'
+    oldMLABMongoURI: 'mongodb://theo:tsoba405012@ds145750.mlab.com:45750/dev_connector',
+    mongoURI: 'mongodb+srv://theo:tsoba405012@dev-connector.bytbr.mongodb.net/dev_connector?retryWrites=true&w=majority',
+    secretOrKey: 'whatever'
 };
