@@ -10,17 +10,17 @@ opts.jwtFromRequest = ExtractJwt.fromAuthHeaderAsBearerToken();
 opts.secretOrKey = keys.secretOrKey;
 
 module.exports = passport => {
-	passport.use(
-		new JwtStrategy(opts, (jwt_payload, done) => {
-			// Find the user from Mongo with id and done is a function returning the user on succ and false on fail
-			User.findById(jwt_payload.id)
-				.then(user => {
-					if (user) {
-						return done(null, user);
-					}
-					return done(null, false);
-				})
-				.catch(err => console.log(err));
-		})
-	);
+    passport.use(
+        new JwtStrategy(opts, (jwt_payload, done) => {
+            // Find the user from Mongo with id and done is a function returning the user on succ and false on fail
+            User.findById(jwt_payload.id)
+                .then(user => {
+                    if (user) {
+                        return done(null, user);
+                    }
+                    return done(null, false);
+                })
+                .catch(err => console.log(err));
+        })
+    );
 };
