@@ -3,25 +3,25 @@ const Schema = mongoose.Schema;
 
 //Create Schema
 const UserSchema = new Schema({
-	name: {
-		type: String,
-		required: true
-	},
-	email: {
-		type: String,
-		required: true
-	},
-	password: {
-		type: String,
-		required: true
-	},
-	avatar: {
-		type: String
-	},
-	date: {
-		type: Date,
-		default: Date.now
-	}
+    name: {
+        type: String,
+        required: true
+    },
+    email: {
+        type: String,
+        required: true
+    },
+    password: {
+        type: String,
+        required: true
+    },
+    avatar: {
+        type: String
+    },
+    date: {
+        type: Date,
+        default: Date.now
+    }
 });
 
 module.exports = User = mongoose.model('users', UserSchema);
