@@ -23,306 +23,306 @@ router.get('/test', (req, res) => res.json({ msg: 'profile works' }));
 // @desc    Get current users profile
 // @access  Private
 router.get(
-	'/',
-	passport.authenticate('jwt', { session: false }),
-	(req, res) => {
-		const errors = {};
+    '/',
+    passport.authenticate('jwt', { session: false }),
+    (req, res) => {
+        const errors = {};
 
-		// req.user.id comes from the token
-		Profile.findOne({ user: req.user.id })
-			.populate('user', ['name', 'avatar']) //This is to fill the profile that is returned with information from the user ref
-			.then(profile => {
-				if (!profile) {
-					errors.profile = 'There is no profile for this user';
-					return res.status(404).json(errors);
-				}
-				res.json(profile);
-			})
-			.catch(err => res.status(404).json(err));
-	}
+        // req.user.id comes from the token
+        Profile.findOne({ user: req.user.id })
+            .populate('user', ['name', 'avatar']) //This is to fill the profile that is returned with information from the user ref
+            .then(profile => {
+                if (!profile) {
+                    errors.profile = 'There is no profile for this user';
+                    return res.status(404).json(errors);
+                }
+                res.json(profile);
+            })
+            .catch(err => res.status(404).json(err));
+    }
 );
 
 // @route   GET api/profile/all
 // @desc    Get all profiles
 // @access  Public
 router.get('/all', (req, res) => {
-	const errors = {};
+    const errors = {};
 
-	Profile.find()
-		.populate('user', ['name', 'avatar']) //This is to fill the profile that is returned with information from the user ref
-		.then(profiles => {
-			if (!profiles) {
-				errors.profile = 'There are no profiles';
-				return res.status(404).json(errors);
-			}
-			res.json(profiles);
-		})
-		.catch(err => {
-			errors.profile = 'There are no profiles';
-			res.status(404).json(errors);
-		});
+    Profile.find()
+        .populate('user', ['name', 'avatar']) //This is to fill the profile that is returned with information from the user ref
+        .then(profiles => {
+            if (!profiles) {
+                errors.profile = 'There are no profiles';
+                return res.status(404).json(errors);
+            }
+            res.json(profiles);
+        })
+        .catch(err => {
+            errors.profile = 'There are no profiles';
+            res.status(404).json(errors);
+        });
 });
 
 // @route   GET api/profile/handle/:handle
 // @desc    Get profile by handle
 // @access  Public
 router.get('/handle/:handle', (req, res) => {
-	const errors = {};
+    const errors = {};
 
-	Profile.findOne({ handle: req.params.handle })
-		.populate('user', ['name', 'avatar']) //This is to fill the profile that is returned with information from the user ref
-		.then(profile => {
-			if (!profile) {
-				errors.profile = 'There is no profile for this user';
-				return res.status(404).json(errors);
-			}
-			res.json(profile);
-		})
-		.catch(err => res.status(404).json(err));
+    Profile.findOne({ handle: req.params.handle })
+        .populate('user', ['name', 'avatar']) //This is to fill the profile that is returned with information from the user ref
+        .then(profile => {
+            if (!profile) {
+                errors.profile = 'There is no profile for this user';
+                return res.status(404).json(errors);
+            }
+            res.json(profile);
+        })
+        .catch(err => res.status(404).json(err));
 });
 
 // @route   GET api/profile/user/:user_id
 // @desc    Get profile by user ID
 // @access  Public
 router.get('/user/:user_id', (req, res) => {
-	const errors = {};
+    const errors = {};
 
-	Profile.findOne({ user: req.params.user_id })
-		.populate('user', ['name', 'avatar']) //This is to fill the profile that is returned with information from the user ref
-		.then(profile => {
-			if (!profile) {
-				errors.profile = 'There is no profile for this user';
-				return res.status(404).json(errors);
-			}
-			res.json(profile);
-		})
-		.catch(err => {
-			errors.profile = 'There is no profile for this user';
-			res.status(404).json(errors);
-		});
+    Profile.findOne({ user: req.params.user_id })
+        .populate('user', ['name', 'avatar']) //This is to fill the profile that is returned with information from the user ref
+        .then(profile => {
+            if (!profile) {
+                errors.profile = 'There is no profile for this user';
+                return res.status(404).json(errors);
+            }
+            res.json(profile);
+        })
+        .catch(err => {
+            errors.profile = 'There is no profile for this user';
+            res.status(404).json(errors);
+        });
 });
 
 // @route   POST api/profile/
 // @desc    Create / Update user profile
 // @access  Private
 router.post(
-	'/',
-	passport.authenticate('jwt', { session: false }),
-	(req, res) => {
-		const { errors, isValid } = validateProfileInput(req.body);
+    '/',
+    passport.authenticate('jwt', { session: false }),
+    (req, res) => {
+        const { errors, isValid } = validateProfileInput(req.body);
 
-		// Check Validation
-		if (!isValid) {
-			// Return any errors with 400 status
-			return res.status(400).json(errors);
-		}
+        // Check Validation
+        if (!isValid) {
+            // Return any errors with 400 status
+            return res.status(400).json(errors);
+        }
 
-		// Check if urls (website and social) contain the http
-		// prefix since validation does not check for it
-		const containsHttp = url => {
-			if (url.includes('http')) {
-				return url;
-			} else {
-				return `http://${url}`;
-			}
-		};
+        // Check if urls (website and social) contain the http
+        // prefix since validation does not check for it
+        const containsHttp = url => {
+            if (url.includes('http')) {
+                return url;
+            } else {
+                return `http://${url}`;
+            }
+        };
 
-		const containsHttps = url => {
-			if (url.includes('https')) {
-				return url;
-			} else {
-				return `https://${url}`;
-			}
-		};
+        const containsHttps = url => {
+            if (url.includes('https')) {
+                return url;
+            } else {
+                return `https://${url}`;
+            }
+        };
 
-		// Get fields
-		const profileFields = {};
-		profileFields.user = req.user.id;
+        // Get fields
+        const profileFields = {};
+        profileFields.user = req.user.id;
 
-		if (req.body.handle) profileFields.handle = req.body.handle;
-		if (req.body.company) profileFields.company = req.body.company;
-		if (req.body.location) profileFields.location = req.body.location;
-		if (req.body.bio) profileFields.bio = req.body.bio;
-		if (req.body.status) profileFields.status = req.body.status;
-		if (req.body.website)
-			profileFields.website = containsHttp(req.body.website);
-		if (req.body.githubusername) {
-			profileFields.githubusername = req.body.githubusername;
-		}
-		// Skills - Split into array
-		if (typeof req.body.skills !== 'undefined') {
-			profileFields.skills = req.body.skills.split(',');
-			profileFields.skills = profileFields.skills.map(str => str.trim());
-		}
+        if (req.body.handle) profileFields.handle = req.body.handle;
+        if (req.body.company) profileFields.company = req.body.company;
+        if (req.body.location) profileFields.location = req.body.location;
+        if (req.body.bio) profileFields.bio = req.body.bio;
+        if (req.body.status) profileFields.status = req.body.status;
+        if (req.body.website)
+            profileFields.website = containsHttp(req.body.website);
+        if (req.body.githubusername) {
+            profileFields.githubusername = req.body.githubusername;
+        }
+        // Skills - Split into array
+        if (typeof req.body.skills !== 'undefined') {
+            profileFields.skills = req.body.skills.split(',');
+            profileFields.skills = profileFields.skills.map(str => str.trim());
+        }
 
-		// Social initialization
-		profileFields.social = {};
-		if (req.body.youtube)
-			profileFields.social.youtube = containsHttps(req.body.youtube);
-		if (req.body.twitter)
-			profileFields.social.twitter = containsHttps(req.body.twitter);
-		if (req.body.facebook)
-			profileFields.social.facebook = containsHttps(req.body.facebook);
-		if (req.body.linkedin)
-			profileFields.social.linkedin = containsHttps(req.body.linkedin);
-		if (req.body.instagram)
-			profileFields.social.instagram = containsHttps(req.body.instagram);
+        // Social initialization
+        profileFields.social = {};
+        if (req.body.youtube)
+            profileFields.social.youtube = containsHttps(req.body.youtube);
+        if (req.body.twitter)
+            profileFields.social.twitter = containsHttps(req.body.twitter);
+        if (req.body.facebook)
+            profileFields.social.facebook = containsHttps(req.body.facebook);
+        if (req.body.linkedin)
+            profileFields.social.linkedin = containsHttps(req.body.linkedin);
+        if (req.body.instagram)
+            profileFields.social.instagram = containsHttps(req.body.instagram);
 
-		Profile.findOne({ user: req.user.id }).then(profile => {
-			if (profile) {
-				// Update
-				Profile.findOneAndUpdate(
-					{ user: req.user.id },
-					{ $set: profileFields },
-					{ new: true }
-				).then(profile => res.json(profile));
-			} else {
-				// Create and check if handle exists
-				Profile.findOne({ handle: profileFields.handle }).then(profile => {
-					if (profile) {
-						errors.handle = 'That handle already exists';
-						res.status(400).json(errors);
-					}
+        Profile.findOne({ user: req.user.id }).then(profile => {
+            if (profile) {
+                // Update
+                Profile.findOneAndUpdate(
+                    { user: req.user.id },
+                    { $set: profileFields },
+                    { new: true }
+                ).then(profile => res.json(profile));
+            } else {
+                // Create and check if handle exists
+                Profile.findOne({ handle: profileFields.handle }).then(profile => {
+                    if (profile) {
+                        errors.handle = 'That handle already exists';
+                        res.status(400).json(errors);
+                    }
 
-					// Save Profile
-					new Profile(profileFields).save().then(profile => res.json(profile));
-				});
-			}
-		});
-	}
+                    // Save Profile
+                    new Profile(profileFields).save().then(profile => res.json(profile));
+                });
+            }
+        });
+    }
 );
 
 // @route   POST api/profile/experience
 // @desc    Add experience to profile
 // @access  Private
 router.post(
-	'/experience',
-	passport.authenticate('jwt', { session: false }),
-	(req, res) => {
-		const { errors, isValid } = validateExperienceInput(req.body);
+    '/experience',
+    passport.authenticate('jwt', { session: false }),
+    (req, res) => {
+        const { errors, isValid } = validateExperienceInput(req.body);
 
-		// Check Validation
-		if (!isValid) {
-			// Return any errors with 400 status
-			return res.status(400).json(errors);
-		}
+        // Check Validation
+        if (!isValid) {
+            // Return any errors with 400 status
+            return res.status(400).json(errors);
+        }
 
-		Profile.findOne({ user: req.user.id }).then(profile => {
-			const newExp = {
-				title: req.body.title,
-				company: req.body.company,
-				location: req.body.location,
-				from: req.body.from,
-				to: req.body.to,
-				current: req.body.current,
-				description: req.body.description
-			};
+        Profile.findOne({ user: req.user.id }).then(profile => {
+            const newExp = {
+                title: req.body.title,
+                company: req.body.company,
+                location: req.body.location,
+                from: req.body.from,
+                to: req.body.to,
+                current: req.body.current,
+                description: req.body.description
+            };
 
-			// Add to experience array (in the beginning with unshift)
-			profile.experience.unshift(newExp);
+            // Add to experience array (in the beginning with unshift)
+            profile.experience.unshift(newExp);
 
-			profile.save().then(profile => res.json(profile));
-		});
-	}
+            profile.save().then(profile => res.json(profile));
+        });
+    }
 );
 
 // @route   POST api/profile/education
 // @desc    Add education to profile
 // @access  Private
 router.post(
-	'/education',
-	passport.authenticate('jwt', { session: false }),
-	(req, res) => {
-		const { errors, isValid } = validateEducationInput(req.body);
+    '/education',
+    passport.authenticate('jwt', { session: false }),
+    (req, res) => {
+        const { errors, isValid } = validateEducationInput(req.body);
 
-		// Check Validation
-		if (!isValid) {
-			// Return any errors with 400 status
-			return res.status(400).json(errors);
-		}
+        // Check Validation
+        if (!isValid) {
+            // Return any errors with 400 status
+            return res.status(400).json(errors);
+        }
 
-		Profile.findOne({ user: req.user.id }).then(profile => {
-			const newEdu = {
-				school: req.body.school,
-				degree: req.body.degree,
-				fieldofstudy: req.body.fieldofstudy,
-				from: req.body.from,
-				to: req.body.to,
-				current: req.body.current,
-				description: req.body.description
-			};
+        Profile.findOne({ user: req.user.id }).then(profile => {
+            const newEdu = {
+                school: req.body.school,
+                degree: req.body.degree,
+                fieldofstudy: req.body.fieldofstudy,
+                from: req.body.from,
+                to: req.body.to,
+                current: req.body.current,
+                description: req.body.description
+            };
 
-			// Add to education array (in the beginning with unshift)
-			profile.education.unshift(newEdu);
+            // Add to education array (in the beginning with unshift)
+            profile.education.unshift(newEdu);
 
-			profile.save().then(profile => res.json(profile));
-		});
-	}
+            profile.save().then(profile => res.json(profile));
+        });
+    }
 );
 
 // @route   DELETE api/profile/experience/:exp_id
 // @desc    Delete experience from profile
 // @access  Private
 router.delete(
-	'/experience/:exp_id',
-	passport.authenticate('jwt', { session: false }),
-	(req, res) => {
-		Profile.findOne({ user: req.user.id }).then(profile => {
-			// Get remove index
-			const removeIndex = profile.experience
-				.map(item => item.id)
-				.indexOf(req.params.exp_id);
+    '/experience/:exp_id',
+    passport.authenticate('jwt', { session: false }),
+    (req, res) => {
+        Profile.findOne({ user: req.user.id }).then(profile => {
+            // Get remove index
+            const removeIndex = profile.experience
+                .map(item => item.id)
+                .indexOf(req.params.exp_id);
 
-			// Splice that out of the array
-			profile.experience.splice(removeIndex, 1);
+            // Splice that out of the array
+            profile.experience.splice(removeIndex, 1);
 
-			// Save
-			profile
-				.save()
-				.then(profile => res.json(profile))
-				.catch(err => res.status(404).json(err));
-		});
-	}
+            // Save
+            profile
+                .save()
+                .then(profile => res.json(profile))
+                .catch(err => res.status(404).json(err));
+        });
+    }
 );
 
 // @route   DELETE api/profile/education/:edu_id
 // @desc    Delete education from profile
 // @access  Private
 router.delete(
-	'/education/:edu_id',
-	passport.authenticate('jwt', { session: false }),
-	(req, res) => {
-		Profile.findOne({ user: req.user.id }).then(profile => {
-			// Get remove index
-			const removeIndex = profile.education
-				.map(item => item.id)
-				.indexOf(req.params.edu_id);
+    '/education/:edu_id',
+    passport.authenticate('jwt', { session: false }),
+    (req, res) => {
+        Profile.findOne({ user: req.user.id }).then(profile => {
+            // Get remove index
+            const removeIndex = profile.education
+                .map(item => item.id)
+                .indexOf(req.params.edu_id);
 
-			// Splice that out of the array
-			profile.education.splice(removeIndex, 1);
+            // Splice that out of the array
+            profile.education.splice(removeIndex, 1);
 
-			// Save
-			profile
-				.save()
-				.then(profile => res.json(profile))
-				.catch(err => res.status(404).json(err));
-		});
-	}
+            // Save
+            profile
+                .save()
+                .then(profile => res.json(profile))
+                .catch(err => res.status(404).json(err));
+        });
+    }
 );
 
 // @route   DELETE api/profile/
 // @desc    Delete user and profile
 // @access  Private
 router.delete(
-	'/',
-	passport.authenticate('jwt', { session: false }),
-	(req, res) => {
-		Profile.findOneAndRemove({ user: req.user.id }).then(() => {
-			User.findOneAndRemove({ _id: req.user.id }).then(() => {
-				res.json({ success: true });
-			});
-		});
-	}
+    '/',
+    passport.authenticate('jwt', { session: false }),
+    (req, res) => {
+        Profile.findOneAndRemove({ user: req.user.id }).then(() => {
+            User.findOneAndRemove({ _id: req.user.id }).then(() => {
+                res.json({ success: true });
+            });
+        });
+    }
 );
 
 module.exports = router;
