@@ -24,113 +24,113 @@ router.get('/test', (req, res) => res.json({ msg: 'users works' }));
 // @desc    Register user
 // @access  Public
 router.post('/register', (req, res) => {
-	const { errors, isValid } = validateRegisterInput(req.body); // Destructuring
+    const { errors, isValid } = validateRegisterInput(req.body); // Destructuring
 
-	// Check Validation
-	if (!isValid) {
-		return res.status(400).json(errors);
-	}
+    // Check Validation
+    if (!isValid) {
+        return res.status(400).json(errors);
+    }
 
-	User.findOne({ email: req.body.email }).then(user => {
-		if (user) {
-			errors.email = 'Email already exists';
-			return res.status(400).json(errors);
-		} else {
-			const avatar = gravatar.url(req.body.email, {
-				s: '200', // Size
-				r: 'pg', // Rating
-				d: 'mm' // Default pic
-			});
+    User.findOne({ email: req.body.email }).then(user => {
+        if (user) {
+            errors.email = 'Email already exists';
+            return res.status(400).json(errors);
+        } else {
+            const avatar = gravatar.url(req.body.email, {
+                s: '200', // Size
+                r: 'pg', // Rating
+                d: 'mm' // Default pic
+            });
 
-			const newUser = new User({
-				name: req.body.name,
-				email: req.body.email,
-				avatar,
-				password: req.body.password
-			});
+            const newUser = new User({
+                name: req.body.name,
+                email: req.body.email,
+                avatar,
+                password: req.body.password
+            });
 
-			// Generate the salt
-			bcrypt.genSalt(10, (err, salt) => {
-				// Encrypt the password using the salt
-				bcrypt.hash(newUser.password, salt, (err, hash) => {
-					if (err) {
-						throw err;
-					}
+            // Generate the salt
+            bcrypt.genSalt(10, (err, salt) => {
+                // Encrypt the password using the salt
+                bcrypt.hash(newUser.password, salt, (err, hash) => {
+                    if (err) {
+                        throw err;
+                    }
 
-					newUser.password = hash;
-					newUser
-						.save()
-						.then(user => {
-							res.json(user);
-						})
-						.catch(err => console.log(err));
-				});
-			});
-		}
-	});
+                    newUser.password = hash;
+                    newUser
+                        .save()
+                        .then(user => {
+                            res.json(user);
+                        })
+                        .catch(err => console.log(err));
+                });
+            });
+        }
+    });
 });
 
 // @route   POST api/users/login
 // @desc    Login User / Returning JWT Token
 // @access  Public
 router.post('/login', (req, res) => {
-	const { errors, isValid } = validateLoginInput(req.body); // Destructuring
+    const { errors, isValid } = validateLoginInput(req.body); // Destructuring
 
-	// Check Validation
-	if (!isValid) {
-		return res.status(400).json(errors);
-	}
+    // Check Validation
+    if (!isValid) {
+        return res.status(400).json(errors);
+    }
 
-	const email = req.body.email;
-	const password = req.body.password;
+    const email = req.body.email;
+    const password = req.body.password;
 
-	// Find  user by email
-	User.findOne({ email }).then(user => {
-		// Check for user
-		if (!user) {
-			errors.email = 'User not found';
-			return res.status(404).json(errors);
-		}
+    // Find  user by email
+    User.findOne({ email }).then(user => {
+        // Check for user
+        if (!user) {
+            errors.email = 'User not found';
+            return res.status(404).json(errors);
+        }
 
-		// Check Password
-		bcrypt.compare(password, user.password).then(isMatched => {
-			if (isMatched) {
-				// User Matched
-				const payload = { id: user.id, name: user.name, avatar: user.avatar }; // Create JWT payload
+        // Check Password
+        bcrypt.compare(password, user.password).then(isMatched => {
+            if (isMatched) {
+                // User Matched
+                const payload = { id: user.id, name: user.name, avatar: user.avatar }; // Create JWT payload
 
-				// Sign Token: payload(user info), secret key, expiration
-				jwt.sign(
-					payload,
-					keys.secretOrKey,
-					{ expiresIn: 3600 },
-					(err, token) => {
-						res.json({
-							success: true,
-							token: `Bearer ${token}`
-						});
-					}
-				);
-			} else {
-				errors.password = 'Password incorrect';
-				return res.status(400).json(errors);
-			}
-		});
-	});
+                // Sign Token: payload(user info), secret key, expiration
+                jwt.sign(
+                    payload,
+                    keys.secretOrKey,
+                    { expiresIn: 3600 },
+                    (err, token) => {
+                        res.json({
+                            success: true,
+                            token: `Bearer ${token}`
+                        });
+                    }
+                );
+            } else {
+                errors.password = 'Password incorrect';
+                return res.status(400).json(errors);
+            }
+        });
+    });
 });
 
 // @route   GET api/users/current
 // @desc    Return current user
 // @access  Private
 router.get(
-	'/current',
-	passport.authenticate('jwt', { session: false }),
-	(req, res) => {
-		res.json({
-			id: req.user.id,
-			name: req.user.name,
-			email: req.user.email
-		});
-	}
+    '/current',
+    passport.authenticate('jwt', { session: false }),
+    (req, res) => {
+        res.json({
+            id: req.user.id,
+            name: req.user.name,
+            email: req.user.email
+        });
+    }
 );
 
 module.exports = router;
