@@ -2,32 +2,32 @@ const Validator = require('validator');
 const isEmpty = require('./is-empty');
 
 module.exports = function validateExperienceInput(data) {
-	let errors = {};
+    let errors = {};
 
-	// If it doesn't pass our custom validation then we make it an empty string in order to fail later
-	data.school = !isEmpty(data.school) ? data.school : '';
-	data.degree = !isEmpty(data.degree) ? data.degree : '';
-	data.fieldofstudy = !isEmpty(data.fieldofstudy) ? data.fieldofstudy : '';
-	data.from = !isEmpty(data.from) ? data.from : '';
+    // If it doesn't pass our custom validation then we make it an empty string in order to fail later
+    data.school = !isEmpty(data.school) ? data.school : '';
+    data.degree = !isEmpty(data.degree) ? data.degree : '';
+    data.fieldofstudy = !isEmpty(data.fieldofstudy) ? data.fieldofstudy : '';
+    data.from = !isEmpty(data.from) ? data.from : '';
 
-	if (Validator.isEmpty(data.school)) {
-		errors.school = 'School field is required';
-	}
+    if (Validator.isEmpty(data.school)) {
+        errors.school = 'School field is required';
+    }
 
-	if (Validator.isEmpty(data.degree)) {
-		errors.degree = 'Degree field is required';
-	}
+    if (Validator.isEmpty(data.degree)) {
+        errors.degree = 'Degree field is required';
+    }
 
-	if (Validator.isEmpty(data.fieldofstudy)) {
-		errors.fieldofstudy = 'Field of study field is required';
-	}
+    if (Validator.isEmpty(data.fieldofstudy)) {
+        errors.fieldofstudy = 'Field of study field is required';
+    }
 
-	if (Validator.isEmpty(data.from)) {
-		errors.from = 'From date field is required';
-	}
+    if (Validator.isEmpty(data.from)) {
+        errors.from = 'From date field is required';
+    }
 
-	return {
-		errors,
-		isValid: isEmpty(errors)
-	};
+    return {
+        errors,
+        isValid: isEmpty(errors)
+    };
 };
