@@ -2,26 +2,26 @@ const Validator = require('validator');
 const isEmpty = require('./is-empty');
 
 module.exports = function validateLoginInput(data) {
-	let errors = {};
+    let errors = {};
 
-	// If it doesn't pass our custom validation then we make it an empty string in order to fail later
-	data.email = !isEmpty(data.email) ? data.email : '';
-	data.password = !isEmpty(data.password) ? data.password : '';
+    // If it doesn't pass our custom validation then we make it an empty string in order to fail later
+    data.email = !isEmpty(data.email) ? data.email : '';
+    data.password = !isEmpty(data.password) ? data.password : '';
 
-	if (!Validator.isEmail(data.email)) {
-		errors.email = 'Email is invalid';
-	}
+    if (!Validator.isEmail(data.email)) {
+        errors.email = 'Email is invalid';
+    }
 
-	if (Validator.isEmpty(data.email)) {
-		errors.email = 'Email field is required';
-	}
+    if (Validator.isEmpty(data.email)) {
+        errors.email = 'Email field is required';
+    }
 
-	if (Validator.isEmpty(data.password)) {
-		errors.password = 'Password field is required';
-	}
+    if (Validator.isEmpty(data.password)) {
+        errors.password = 'Password field is required';
+    }
 
-	return {
-		errors,
-		isValid: isEmpty(errors)
-	};
+    return {
+        errors,
+        isValid: isEmpty(errors)
+    };
 };
